feat: removal of the HideClient module
Because people kept "accidentally" clicking on it and not knowing how to turn it off, we decided to remove the module.
