+++ src/main/kotlin/net/ccbluex/liquidbounce/features/misc/HideAppearance.kt
@@ -1,7 +1,7 @@
 /*
  * This file is part of LiquidBounce (https://github.com/CCBlueX/LiquidBounce)
  *
- * Copyright (c) 2015-2024 CCBlueX
+ * Copyright (c) 2015 - 2024 CCBlueX
  *
  * LiquidBounce is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,44 +15,46 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with LiquidBounce. If not, see <https://www.gnu.org/licenses/>.
- *
- *
  */
-package net.ccbluex.liquidbounce.features.module.modules.client
+package net.ccbluex.liquidbounce.features.misc
 
+import net.ccbluex.liquidbounce.event.Listenable
 import net.ccbluex.liquidbounce.event.events.KeyboardKeyEvent
 import net.ccbluex.liquidbounce.event.handler
-import net.ccbluex.liquidbounce.features.misc.HideClient
-import net.ccbluex.liquidbounce.features.module.Category
-import net.ccbluex.liquidbounce.features.module.Module
 import net.ccbluex.liquidbounce.utils.client.Chronometer
 import net.ccbluex.liquidbounce.utils.client.inGame
+import net.ccbluex.liquidbounce.utils.client.mc
 import net.ccbluex.liquidbounce.web.integration.IntegrationHandler
+import net.minecraft.SharedConstants
+import net.minecraft.client.util.Icons
 import org.lwjgl.glfw.GLFW
 
 /**
- * Hides client
+ * Hides client appearance
+ *
+ * using 2x CRTL + SHIFT to hide and unhide the client
  */
-object ModuleHideClient : Module("HideClient", Category.CLIENT, hide = true) {
+object HideAppearance : Listenable {
 
     val shiftChronometer = Chronometer()
 
-    override fun enable() {
-        HideClient.isHidingNow = true
-
-        // This will cause a refresh of the current screen
-        IntegrationHandler.restoreOriginalScreen()
-        super.enable()
-    }
-
-    override fun disable() {
-        HideClient.isHidingNow = false
-
-        // This will cause a refresh of the current screen
-        if (mc.currentScreen != null) {
-            mc.setScreen(mc.currentScreen)
+    var isHidingNow = false
+        set(value) {
+            field = value
+            updateClient()
         }
-        super.disable()
+
+    private fun updateClient() {
+        mc.updateWindowTitle()
+        mc.window.setIcon(
+            mc.defaultResourcePack,
+            if (SharedConstants.getGameVersion().isStable) Icons.RELEASE else Icons.SNAPSHOT)
+
+        if (isHidingNow) {
+            IntegrationHandler.restoreOriginalScreen()
+        } else {
+            IntegrationHandler.updateIntegrationBrowser()
+        }
     }
 
     val keyHandler = handler<KeyboardKeyEvent>(ignoreCondition = true) {
@@ -65,15 +67,7 @@ object ModuleHideClient : Module("HideClient", Category.CLIENT, hide = true) {
 
         if (keyCode == GLFW.GLFW_KEY_LEFT_SHIFT && modifier == GLFW.GLFW_MOD_CONTROL) {
             if (!shiftChronometer.hasElapsed(400L)) {
-                enabled = !enabled
-
-                // Since we are not in a game, we have to manually enable/disable the module
-                if (enabled) {
-                    enable()
-                } else {
-                    disable()
-                    IntegrationHandler.updateIntegrationBrowser()
-                }
+                isHidingNow = !isHidingNow
             }
 
             shiftChronometer.reset()
